Added base JSX components - SafeView & ScrollView
