+++ frontend/components/MapScreen.js
@@ -1,0 +1,49 @@
+import React from 'react';
+import {
+  SafeAreaView,
+  StyleSheet,
+  ScrollView,
+  View,
+  Button,
+  Text,
+  Platform,
+  StatusBar,
+} from 'react-native';
+import Colors from "./Colors";
+
+
+
+export default function MapScreen({ routes, navigation }) {
+
+  return (
+    <>
+      <SafeAreaView>
+        <ScrollView
+          contentInsetAdjustmentBehavior="automatic"
+          style={styles.scrollView}>
+
+          <View style={styles.body}>
+            <Text>This is the Map</Text>
+          </View>
+        </ScrollView>
+      </SafeAreaView>
+    </>
+  );
+}
+
+
+const styles = StyleSheet.create({
+  scrollView: {
+    backgroundColor: Colors.white,
+  },
+  engine: {
+    position: 'absolute',
+    right: 0,
+  },
+  body: {
+    backgroundColor: Colors.white,
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+});
