--- frontend/containers/StepsScreen.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import {
-  SafeAreaView,
-  StyleSheet,
-  ScrollView,
-  View,
-  Button,
-  Text,
-  StatusBar,
-} from 'react-native';
-
-export default function StepsScreen({ routes, navigation }) {
-
-    return(
-        <>
-        <Text>This is the Steps</Text>
-        </>
-    );
-}
